Fixing bugs

+++ data/ratings_log.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Float, ForeignKey, Integer
+from sqlalchemy import Column, ForeignKey, Integer
 from sqlalchemy.orm import relationship
 from .db_session import SqlAlchemyBase
 
@@ -9,7 +9,7 @@ class RatingLog(SqlAlchemyBase):
     id = Column(Integer, primary_key=True, autoincrement=True)
     dish_id = Column(Integer, ForeignKey("dishes.id"))
     student_id = Column(Integer, ForeignKey("users.id"))
-    rate = Column(Float, nullable=True)
+    rate = Column(Integer, nullable=True)
 
     dishes = relationship("Dish", back_populates="ratings_log")
     users = relationship("User", back_populates="ratings_log")
